% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasto_mensual.R
\name{generar_gasto_mensual}
\alias{generar_gasto_mensual}
\title{Generar gasto mensual en soles}
\usage{
generar_gasto_mensual(n)
}
\arguments{
\item{n}{Número de datos de gasto a generar.}
}
\value{
Vector de gastos mensuales aleatorios entre 200 y 3000 soles.
}
\description{
Esta función genera valores aleatorios simulando el gasto mensual de personas en soles.
}
\examples{
generar_gasto_mensual(10)
}
