% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notas.R
\name{generar_notas}
\alias{generar_notas}
\title{Generar notas aleatorias de estudiantes}
\usage{
generar_notas(n)
}
\arguments{
\item{n}{Número de notas a generar.}
}
\value{
Vector con notas aleatorias.
}
\description{
Esta función genera un vector de notas aleatorias enteras entre 0 y 20.
}
\examples{
generar_notas(10)
}
