% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tallas.R
\name{generar_tallas}
\alias{generar_tallas}
\title{Generar tallas aleatorias}
\usage{
generar_tallas(n)
}
\arguments{
\item{n}{Número de tallas a generar.}
}
\value{
Vector de tallas aleatorias.
}
\description{
Esta función genera un vector de tallas aleatorias entre 0.5 y 1.99.
}
\examples{
generar_tallas(10)
}
