% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperaturas.R
\name{generar_temperaturas}
\alias{generar_temperaturas}
\title{Generar temperaturas ambientales aleatorias}
\usage{
generar_temperaturas(n)
}
\arguments{
\item{n}{Número de temperaturas a generar.}
}
\value{
Vector de temperaturas aleatorias con una precisión de una decimal.
}
\description{
Esta función genera temperaturas aleatorias entre -10°C y 40°C, simulando datos ambientales.
}
\examples{
generar_temperaturas(10)
}
